#48 - create recipe

+++ src/app/recipes/ui/components/recipe-components/recipe-detail/recipe-detail.component.ts
@@ -3,7 +3,7 @@ import { IonGrid, IonRow, IonCol, IonSearchbar, IonButton, IonIcon, IonAlert, Io
 import { VariantsListComponent } from '../../recipe-variant-components/variants-list/variants-list.component';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { RecipeVariantDetailComponent } from '../../recipe-variant-components/recipe-variant-detail/recipe-variant-detail.component';
-import { Recipe } from 'src/app/recipes/data/interfaces/recipe.interface';
+import { IRecipe } from 'src/app/recipes/data/interfaces/recipe.interface';
 import { Course } from 'src/app/recipes/data/enums/courses.enum';
 import { CommonModule } from '@angular/common';
 import { TranslateStore } from "@ngx-translate/core";
@@ -11,6 +11,7 @@ import { addIcons } from 'ionicons';
 import { add, trash } from 'ionicons/icons';
 import { RecipesService } from 'src/app/recipes/data/services/recipes.service';
 import { CourseListComponent } from '../course-list/course-list.component';
+import { AlertService } from '../../../services/alert.service';
 
 @Component({
   selector: 'app-recipe-detail',
@@ -18,7 +19,6 @@ import { CourseListComponent } from '../course-list/course-list.component';
   styleUrls: ['./recipe-detail.component.scss'],
   standalone: true,
   imports: [IonPopover, IonIcon, IonButton,
-    IonAlert,
     IonGrid,
     IonRow,
     IonCol,
@@ -32,31 +32,13 @@ import { CourseListComponent } from '../course-list/course-list.component';
 })
 export class RecipeDetailComponent {
 
-  @Input({ required: true }) recipe!: Recipe;
+  @Input({ required: true }) recipe!: IRecipe;
   @Input() isEditing = false;
 
 
   @ViewChild(IonPopover) popover!: IonPopover;
 
   selectedIds: number[] = [];
-
-  alertButtons = [
-    {
-      text: this.translateService.instant('alert.cancel'),
-      role: 'cancel',
-      handler: () => {
-        console.log('Alert cancelled');
-      },
-    },
-    {
-      text: this.translateService.instant('alert.ok'),
-      role: 'confirm',
-      handler: () => {
-        console.log('Alert confirmed');
-        this.deleteSelected();
-      },
-    },
-  ];
 
   get selectedCount(): number {
     return this.selectedIds.length;
@@ -67,7 +49,8 @@ export class RecipeDetailComponent {
   }
 
   constructor(private translateService: TranslateService,
-    private recipesService: RecipesService) {
+    private recipesService: RecipesService,
+    private alertService: AlertService) {
 
     addIcons({ add, trash });
 
@@ -96,15 +79,24 @@ export class RecipeDetailComponent {
   }
 
   deleteSelected() {
-    this.recipesService.deleteVariants(this.recipe.id, this.selectedIds).subscribe({
-      next: () => {
-        this.selectedIds = [];
-      },
-      error: (err: any) => {
-        console.error(err);
-      }
-    }
-    );
+
+    this.alertService.presentConfirm(
+      this.translateService.instant('recipe-detail.delete-variants'),
+      this.translateService.instant('recipe-detail.delete-variants-message'),
+      () => {
+        if (this.recipe.id) {
+          this.recipesService.deleteVariants(this.recipe.id, this.selectedIds).subscribe({
+            next: (variants) => {
+              this.selectedIds = [];
+              this.recipe.variants = variants;
+            },
+            error: (err: any) => {
+              console.error(err);
+            }
+          }
+          );
+        }
+      });
   }
 
   selectedCoursesChange(courses: Course[]) {
