+++ src/app/recipes/ui/pages/recipe-detail.page/recipe-detail.page.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { RecipesService } from '../../../data/services/recipes.service';
 import { IonHeader, IonToolbar, IonContent, IonButtons, IonButton, IonIcon, IonBackButton } from '@ionic/angular/standalone';
-import { Recipe } from '../../../data/interfaces/recipe.interface';
+import { IRecipe } from '../../../data/interfaces/recipe.interface';
 import { CommonModule } from '@angular/common';
 import { addIcons } from 'ionicons';
 import { pencil, trash } from 'ionicons/icons';
@@ -31,7 +31,7 @@ import { AlertService } from '../../services/alert.service';
 })
 export class RecipeDetailPage implements OnInit {
 
-  recipe: Recipe | null = null;
+  recipe: IRecipe | null = null;
   component = RecipesPage;
 
   constructor(
@@ -50,7 +50,7 @@ export class RecipeDetailPage implements OnInit {
   ngOnInit() {
     this.route.params.subscribe(params => {
       const recipeId = +params['id'];
-      this.recipesService.getRecipe(recipeId).subscribe((recipe: Recipe) => {
+      this.recipesService.getRecipe(recipeId).subscribe((recipe: IRecipe) => {
         this.recipe = recipe;
       });
     });
@@ -78,7 +78,7 @@ export class RecipeDetailPage implements OnInit {
       this.translateService.instant('alert.delete-confirm'),
       message,
       () => {
-        if (this.recipe) {
+        if (this.recipe && this.recipe.id) {
           this.recipesService.deleteRecipe(this.recipe.id).subscribe({
             next: async () => {
               this.router.navigate(['/tabs/recipes']);
